removed `numpy` dependency that was unused an triggering pyo3 dependency in downstream packages various tidying up steps

+++ fastsim-core/src/vehicle/vehicle_model.rs
@@ -31,7 +31,7 @@ pub struct Vehicle {
     /// Vehicle name
     name: String,
     /// Year manufactured
-    year: u32,
+    year: u8,
     #[has_state]
     /// type of vehicle powertrain including contained type-specific parameters and variables
     pub pt_type: PowertrainType,
